Create Nova resources

+++ app/Http/Controllers/ProductController.php
@@ -86,14 +86,14 @@ class ProductController extends Controller
     {
 
         $tour = Tour::findOrFail($id);
-       // $this->authorize('update', $tour);
+        $this->authorize('update', $tour);
         return view('pages.product.product-update.product-update', compact('tour'));
     }
 
     public function update (ProductRequest $request, $id)
     {
         $tour = Tour::findOrFail($id);
-       // $this->authorize('update', $tour);
+        $this->authorize('update', $tour);
 
         $tour = $this->helper->updateTour($request, $tour);
         return redirect()->route('product-page', $tour);
@@ -104,7 +104,7 @@ class ProductController extends Controller
     public function destroy ($id)
     {
         $tour = Tour::find($id);
-       // $this->authorize('update', $tour);
+        $this->authorize('delete', $tour);
         $tour->medias()->detach();
         $tour->delete();
         return redirect('/');
